+++ Packages/DotNET/Configurations/Source/Tenancy/TenantOptionsFactory.cs
@@ -4,6 +4,7 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Options;
 using Woksin.Extensions.Tenancy;
+using Woksin.Extensions.Tenancy.Context;
 
 namespace Woksin.Extensions.Configurations.Tenancy;
