+++ Packages/DotNET/Tenancy/Source/Base/Context/ITenantContextAccessor.cs
@@ -1,7 +1,7 @@
 // Copyright (c) woksin-org. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Woksin.Extensions.Tenancy;
+namespace Woksin.Extensions.Tenancy.Context;
 
 /// <summary>
 /// Defines a system that can access the current <see cref="ITenantContext"/>.
