+++ Packages/DotNET/Tenancy/Source/Base/Context/TenantContextIsNotResolved.cs
@@ -1,7 +1,7 @@
 // Copyright (c) woksin-org. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Woksin.Extensions.Tenancy;
+namespace Woksin.Extensions.Tenancy.Context;
 
 /// <summary>
 /// Exception that gets thrown when the current tenant context is not set.
